requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-money==3.5.3
drf-spectacular==0.27.2